green: moved Ladder-Tests and Methods to a new class LadderMoverTests
